website/website/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agents>=1.4.0",
    "cohere>=5.20.1",
    "fastapi>=0.127.0",
    "openai-sdk>=0.1.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
    "trafilatura>=2.0.0",
]
